pyproject.toml
[tool.poetry]
name = "sdss-lvmecp"
version = "0.4.0-alpha.0"
description = "A library and actor to communicate with an SDSS-V LVM Enclosure"
authors = ["mingyeong yang <mingyeong@khu.ac.kr>", "José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmecp"
repository = "https://github.com/sdss/lvmecp"
documentation = "https://sdss-lvmecp.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmecp", from = "python" }
]
include = ["python/lvmecp/etc/*"]

[tool.poetry.scripts]
lvmecp = "lvmecp.__main__:lvmecp"

[tool.poetry.dependencies]
python = "^3.7"
sdsstools = ">=0.4.0"
click = "^8.0.1"
sdss-clu = "^1.4.0"
click-default-group = "^1.2.2"
pymodbus = "^2.5.3"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
isort = ">=4.3.21"
ipdb = ">=0.12.3"
rstcheck = ">=3.3.1"
black = ">=21.7b0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = ">=4.1.2"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18-beta.36"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"
sphinx-click = ">=3.0.1"

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmecp"
known_sdss = ["sdsstools", "clu", "drift"]
lines_after_imports = 2
profile = "black"

[tool.pytest.ini_options]
addopts = "--cov lvmecp --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
omit = [
    "python/lvmecp/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING"
]

[build-system]
requires = ["poetry-core>=1.1.0a7"]
build-backend = "poetry.core.masonry.api"
